requirements.txt
google-generativeai
google-auth
python-dotenv
Django>=5.0.1
